personal_media_vault/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
pydantic-settings>=2.2.1
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
PyJWT>=2.8.0
imagekitio>=5.2.0
